+++ src/py/_pyodide/_core_docs.py
@@ -108,6 +108,37 @@ class JsProxy:
         an ArrayBuffer view.
         """
 
+    def to_memoryview(self) -> memoryview:
+        """Convert the buffer to a memoryview.
+
+        Copies the data once. This currently has the same effect as :any:`to_py`.
+        Present only if the wrapped Javascript object is an ArrayBuffer or
+        an ArrayBuffer view.
+        """
+
+    def to_bytes(self) -> bytes:
+        """Convert the buffer to a bytes object.
+
+        Copies the data once.
+        Present only if the wrapped Javascript object is an ArrayBuffer or
+        an ArrayBuffer view.
+        """
+
+    def to_string(self, encoding=None) -> str:
+        """Convert the buffer to a string object.
+
+        Copies the data twice.
+
+        The encoding argument will be passed to the Javascript
+        [``TextDecoder``](https://developer.mozilla.org/en-US/docs/Web/API/TextDecoder)
+        constructor. It should be one of the encodings listed in the table here:
+        `https://encoding.spec.whatwg.org/#names-and-labels`. The default
+        encoding is utf8.
+
+        Present only if the wrapped Javascript object is an ArrayBuffer or
+        an ArrayBuffer view.
+        """
+
 
 # from pyproxy.c